pyproject.toml
[tool.poetry]
name = "unistore"
version = "0.1.0"
description = "A single vendor e-commerce site."
authors = ["Aaditya Dulal <artdityadulal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
black = "^21.6b0"
django-crispy-forms = "^1.12.0"
Pillow = "^8.3.1"
django-resized = "^0.3.11"
django-ckeditor = "^6.1.0"
django-tinymce = "^3.3.0"
python-decouple = "^3.4"
django-filter = "^2.4.0"
pre-commit = "^2.14.0"
isort = "^5.9.3"
Sphinx = "^4.1.2"
django-docs = "^0.3.1"
selenium = "^3.141.0"
mock = "^4.0.3"
mixer = "^7.1.2"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"
bandit = "^1.7.0"
pre-commit = "^2.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
